+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## develop
+
+* Enable WAL mode on database open to allow concurrent read/write access across multiple SQLite connections to the same file.
+
 ## 0.7.0
 
 * Initial release.